requirements.txt
Django==4.1.7
petl==1.7.12
requests==2.28.2
python-dateutil==2.8.2
